added mor xml docs
