+++ src/core/Akka.Persistence.TestKit/Journal/TestJournalRejectionException.cs
@@ -1,0 +1,21 @@
+﻿// -----------------------------------------------------------------------
+//  <copyright file="TestJournalRejectionException.cs" company="Akka.NET Project">
+//      Copyright (C) 2009-2019 Lightbend Inc. <http://www.lightbend.com>
+//      Copyright (C) 2013-2019 .NET Foundation <https://github.com/akkadotnet/akka.net>
+//  </copyright>
+// -----------------------------------------------------------------------
+
+namespace Akka.Persistence.TestKit
+{
+    using System;
+    using System.Runtime.Serialization;
+
+    [Serializable]
+    public class TestJournalRejectionException : Exception
+    {
+        public TestJournalRejectionException() { }
+        public TestJournalRejectionException(string message) : base(message) { }
+        public TestJournalRejectionException(string message, Exception inner) : base(message, inner) { }
+        protected TestJournalRejectionException(SerializationInfo info, StreamingContext context) : base(info, context) { }
+    }
+}
