+++ src/core/Akka.Persistence.TestKit/SnapshotStore/SnapshotStoreLoadBehaviorSetter.cs
@@ -11,6 +11,9 @@ namespace Akka.Persistence.TestKit
     using System.Threading.Tasks;
     using Actor;
 
+    /// <summary>
+    ///     Setter strategy for <see cref="TestSnapshotStore"/> which will set load interceptor.
+    /// </summary>
     internal class SnapshotStoreLoadBehaviorSetter : ISnapshotStoreBehaviorSetter
     {
         internal SnapshotStoreLoadBehaviorSetter(IActorRef snapshots)
